// frontend/src/api/generated/models/create-product-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StorageCondition } from './storage-condition';

/**
 * 
 * @export
 * @interface CreateProductRequest
 */
export interface CreateProductRequest {
    /**
     * 商品コード（登録後変更不可）
     * @type {string}
     * @memberof CreateProductRequest
     */
    'productCode': string;
    /**
     * 商品名
     * @type {string}
     * @memberof CreateProductRequest
     */
    'productName': string;
    /**
     * 商品名カナ
     * @type {string}
     * @memberof CreateProductRequest
     */
    'productNameKana'?: string;
    /**
     * ケース入数
     * @type {number}
     * @memberof CreateProductRequest
     */
    'caseQuantity': number;
    /**
     * ボール入数
     * @type {number}
     * @memberof CreateProductRequest
     */
    'ballQuantity': number;
    /**
     * バーコード / JANコード
     * @type {string}
     * @memberof CreateProductRequest
     */
    'barcode'?: string;
    /**
     * 
     * @type {StorageCondition}
     * @memberof CreateProductRequest
     */
    'storageCondition': StorageCondition;
    /**
     * 危険物フラグ
     * @type {boolean}
     * @memberof CreateProductRequest
     */
    'isHazardous': boolean;
    /**
     * ロット管理フラグ
     * @type {boolean}
     * @memberof CreateProductRequest
     */
    'lotManageFlag': boolean;
    /**
     * 賞味/使用期限管理フラグ
     * @type {boolean}
     * @memberof CreateProductRequest
     */
    'expiryManageFlag': boolean;
    /**
     * 出荷禁止フラグ
     * @type {boolean}
     * @memberof CreateProductRequest
     */
    'shipmentStopFlag': boolean;
    /**
     * 有効フラグ
     * @type {boolean}
     * @memberof CreateProductRequest
     */
    'isActive': boolean;
}