// frontend/src/api/generated/models/update-product-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StorageCondition } from './storage-condition';

/**
 * 
 * @export
 * @interface UpdateProductRequest
 */
export interface UpdateProductRequest {
    /**
     * 商品名
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'productName': string;
    /**
     * 商品名カナ
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'productNameKana'?: string;
    /**
     * ケース入数
     * @type {number}
     * @memberof UpdateProductRequest
     */
    'caseQuantity': number;
    /**
     * ボール入数
     * @type {number}
     * @memberof UpdateProductRequest
     */
    'ballQuantity': number;
    /**
     * バーコード / JANコード
     * @type {string}
     * @memberof UpdateProductRequest
     */
    'barcode'?: string;
    /**
     * 
     * @type {StorageCondition}
     * @memberof UpdateProductRequest
     */
    'storageCondition': StorageCondition;
    /**
     * 危険物フラグ
     * @type {boolean}
     * @memberof UpdateProductRequest
     */
    'isHazardous': boolean;
    /**
     * ロット管理フラグ（在庫存在時は変更不可）
     * @type {boolean}
     * @memberof UpdateProductRequest
     */
    'lotManageFlag': boolean;
    /**
     * 賞味/使用期限管理フラグ（在庫存在時は変更不可）
     * @type {boolean}
     * @memberof UpdateProductRequest
     */
    'expiryManageFlag': boolean;
    /**
     * 出荷禁止フラグ
     * @type {boolean}
     * @memberof UpdateProductRequest
     */
    'shipmentStopFlag': boolean;
    /**
     * 有効フラグ
     * @type {boolean}
     * @memberof UpdateProductRequest
     */
    'isActive': boolean;
    /**
     * 楽観的ロックバージョン
     * @type {number}
     * @memberof UpdateProductRequest
     */
    'version': number;
}